test: 결제 관련 테스트 클래스 제거
- `PaymentBoundaryTest`, `PaymentConcurrencyTest`, `PaymentFacadeTest`, `PaymentIdempotencyTest` 삭제
- 통합 테스트 및 단위 테스트 대상 코드 구조 변경에 따른 불필요한 테스트 제거
- 테스트 데이터 초기화 및 Mock 구성 제거

낙후된 테스트 코드 정리 및 최신 테스트 커버리지 유지
